Adapt project for camera triggers
