+++ micNotifyUDP.csproj
@@ -10,6 +10,8 @@
 
   <ItemGroup>
     <PackageReference Include="NAudio" Version="2.2.1" />
+    <PackageReference Include="OpenCvSharp4" Version="4.9.0" />
+    <PackageReference Include="OpenCvSharp4.runtime.win" Version="4.9.0" />
   </ItemGroup>
 
 </Project>
